trips/api/versions/v1_0/serializers/actions.py
from rest_framework import serializers


class DriverActionsSerializer(serializers.Serializer):
    action = serializers.ChoiceField(
        label="Ação a ser realizada no passageiro",
        choices=(('approve', 'Aprovar'),
                 ('deny', 'Negar'), ('forfeit', 'Remover'))
    )
